refactor(web): P4 — carve wave 3 (sessions_chat, agents_loops)
RFC-003 P4. Six sections move verbatim: chat + sessions + trajectories
-> sessions_chat.py; autonomous loops + agents + processes ->
agents_loops.py. __init__ is down to 1,164 lines. The composition
assertions (boundaries stop at call lines; defs == root calls; no
hidden composition) all held first pass this wave.

The interesting fix: the R1 patch seam. The moved chat/execute handlers
bound process_web_chat from their own module, so
patch("src.web.api.process_web_chat") — the pinned historical seam —
stopped governing them (7 test failures caught it immediately).
sessions_chat now resolves the callable through the package attribute
at call time (_pkg_process_web_chat), so the seam promise holds for
real handlers, not just for importability.

Per-registrar admin gates added where the moved sections used the
root's closure (sessions, trajectories). __init__ sheds now-unused
imports; process_web_chat stays re-exported with the seam note.

Gate: new=0. Suite 6,079; parity 183 exact-order.
