+++ src/web/api/__init__.py
@@ -7,7 +7,6 @@ from __future__ import annotations
 
 import asyncio
 import time
-import uuid
 from pathlib import Path
 from typing import TYPE_CHECKING
 
@@ -49,7 +48,12 @@ from ..api_common import (  # noqa: F401 — re-exports
     _write_env_file,
     admin_gate,
 )
-from ..chat import MAX_CHAT_CONTENT_LEN, process_web_chat
+from ..chat import process_web_chat  # noqa: F401 — R1 patch seam + import surface
+from .agents_loops import (
+    register_agents,
+    register_loops,
+    register_processes,
+)
 from .knowledge_mem import (
     register_knowledge,
     register_learned_context,
@@ -84,6 +88,11 @@ from .security import (
     register_auth,
     register_host_access,
     register_permissions_rbac,
+)
+from .sessions_chat import (
+    register_chat,
+    register_sessions,
+    register_trajectories,
 )
 from .skills_api import register_skills
 
@@ -776,349 +785,9 @@ def create_api_routes(bot: OdinBot) -> web.RouteTableDef:
         result = bot.loop_manager.stop_loop("all")
         return web.json_response({"result": result})
 
-    # ------------------------------------------------------------------
-    # Chat
-    # ------------------------------------------------------------------
+    register_chat(routes, bot)
 
-    @routes.post("/api/chat")
-    async def chat(request: web.Request) -> web.Response:
-        try:
-            data = await request.json()
-        except Exception:
-            return web.json_response({"error": "invalid JSON"}, status=400)
-
-        content = (data.get("content") or "").strip()
-        if not content:
-            return web.json_response({"error": "content is required"}, status=400)
-        if len(content) > MAX_CHAT_CONTENT_LEN:
-            return web.json_response(
-                {"error": f"content exceeds {MAX_CHAT_CONTENT_LEN} chars"}, status=400
-            )
-
-        identity = getattr(request, "_api_identity", None)
-        user_id = identity.user_id if identity else "web-user"
-        username = identity.username if identity else "WebUser"
-        tier = identity.tier if identity else None
-        token_tools = identity.allowed_tools if identity and identity.allowed_tools else None
-        token_hosts = (
-            identity.allowed_hosts
-            if identity and isinstance(getattr(identity, "allowed_hosts", None), list)
-            else None
-        )
-        token_default_host = getattr(identity, "default_host", "") if identity else ""
-
-        # Optional caller-supplied session id for multi-request chat continuity.
-        # Omitted -> historical behavior (one history per identity). Supplied -> validated
-        # and namespaced UNDER the authenticated identity. It only controls conversation
-        # continuity + lock serialization; permissions, tier, tools/hosts, memory, and
-        # audit identity all stay keyed to the authenticated token, never the session id.
-        channel_id = user_id
-        session_id = data.get("session_id")
-        if session_id is not None:
-            session_id = session_id.strip() if isinstance(session_id, str) else ""
-            if not _SESSION_ID_RE.match(session_id):
-                return web.json_response(
-                    {"error": "invalid session_id (expected 1-128 chars of [A-Za-z0-9._:-])"},
-                    status=400,
-                )
-            channel_id = _scoped_chat_channel(user_id, session_id)
-
-        result = await process_web_chat(
-            bot, content, channel_id,
-            user_id=user_id, username=username,
-            allowed_tools=token_tools, tier=tier,
-            token_allowed_hosts=token_hosts,
-            token_default_host=token_default_host,
-        )
-
-        # Scoped-session locks are cached like the default per-identity lock. We do NOT
-        # clean them up per-request: that races a waiter and can split one session across
-        # two lock objects (concurrent _do_process_web_chat). Bounding _web_channel_locks
-        # via a TTL/max-size sweep is a deliberate follow-up; correct serialization first.
-        status = 200 if not result["is_error"] else 502
-        resp = {
-            "response": result["response"],
-            "tools_used": result["tools_used"],
-            "is_error": result["is_error"],
-        }
-        if session_id is not None:
-            resp["session_id"] = session_id
-        files = result.get("files", [])
-        if files:
-            resp["files"] = files
-        return web.json_response(resp, status=status)
-
-    @routes.post("/api/execute")
-    async def execute(request: web.Request) -> web.Response:
-        """Stateless prompt execution — no session history, no persistence.
-
-        Designed for CLI tools, scripts, CI/CD pipelines, and automation.
-        Each request gets a unique ephemeral channel_id that is discarded
-        after the response is returned.
-        """
-        try:
-            data = await request.json()
-        except Exception:
-            return web.json_response({"error": "invalid JSON"}, status=400)
-
-        content = (data.get("prompt") or data.get("content") or "").strip()
-        if not content:
-            return web.json_response({"error": "prompt is required"}, status=400)
-        if len(content) > MAX_CHAT_CONTENT_LEN:
-            return web.json_response(
-                {"error": f"prompt exceeds {MAX_CHAT_CONTENT_LEN} chars"}, status=400
-            )
-
-        channel_id = f"api-{uuid.uuid4().hex[:12]}"
-
-        # Resolve identity from middleware or fallback
-        identity = getattr(request, "_api_identity", None)
-        if identity is None:
-            auth_header = request.headers.get("Authorization", "")
-            bearer_token = auth_header[7:] if auth_header.startswith("Bearer ") else ""
-            tm = getattr(bot, "api_token_manager", None)
-            identity = tm.resolve(bearer_token) if tm else None
-            if identity is None:
-                identity = bot.config.web.resolve_api_identity(bearer_token)
-        user_id = identity.user_id if identity else "api-user"
-        username = identity.username if identity else "API"
-        token_tools = identity.allowed_tools if identity and identity.allowed_tools else None
-        tier = identity.tier if identity else None
-        token_hosts = (
-            identity.allowed_hosts
-            if identity and isinstance(getattr(identity, "allowed_hosts", None), list)
-            else None
-        )
-        token_default_host = getattr(identity, "default_host", "") if identity else ""
-
-        result = await process_web_chat(
-            bot, content, channel_id,
-            user_id=user_id, username=username,
-            allowed_tools=token_tools, tier=tier,
-            token_allowed_hosts=token_hosts,
-            token_default_host=token_default_host,
-            persist_channel_lock=False,  # ephemeral per-request channel — no lock to cache or leak
-        )
-
-        bot.sessions.reset(channel_id)
-
-        status = 200 if not result["is_error"] else 502
-        resp = {
-            "response": result["response"],
-            "tools_used": result["tools_used"],
-            "is_error": result["is_error"],
-            "source": "web_api",
-        }
-        if identity and identity.label:
-            resp["token_label"] = identity.label
-        files = result.get("files", [])
-        if files:
-            resp["files"] = files
-        return web.json_response(resp, status=status)
-
-    # ------------------------------------------------------------------
-    # Sessions
-    # ------------------------------------------------------------------
-
-    @routes.get("/api/sessions")
-    async def list_sessions(request: web.Request) -> web.Response:
-        identity = getattr(request, "_api_identity", None)
-        is_admin = not identity or getattr(identity, "tier", "admin") == "admin"
-        own_id = identity.user_id if identity else None
-        sessions = []
-        for cid, session in bot.sessions.items_snapshot():
-            if not is_admin and cid != own_id and not cid.startswith(f"web:{own_id}:session:"):
-                continue
-            # Build preview from last 2 messages
-            preview = []
-            for m in session.messages[-2:]:
-                text = m.content or ""
-                if len(text) > 120:
-                    text = text[:120] + "..."
-                preview.append({"role": m.role, "content": text})
-            # Determine source type
-            source = "web" if cid.startswith(("web-", "web:")) else "discord"
-            sessions.append({
-                "channel_id": cid,
-                "message_count": len(session.messages),
-                "estimated_tokens": session.estimated_tokens,
-                "last_active": session.last_active,
-                "created_at": session.created_at,
-                "has_summary": bool(session.summary),
-                "preview": preview,
-                "source": source,
-                "last_user_id": session.last_user_id,
-            })
-        sessions.sort(key=lambda s: s["last_active"], reverse=True)
-        return web.json_response(sessions)
-
-    @routes.get("/api/sessions/token-usage")
-    async def session_token_usage(request: web.Request) -> web.Response:
-        denied = _require_admin(request)
-        if denied:
-            return denied
-        usage = bot.sessions.get_session_token_usage()
-        return web.json_response(usage)
-
-    @routes.get("/api/sessions/activity")
-    async def session_activity(request: web.Request) -> web.Response:
-        denied = _require_admin(request)
-        if denied:
-            return denied
-        activity = bot.sessions.get_activity_metrics()
-        return web.json_response(activity)
-
-    @routes.get("/api/sessions/search")
-    async def search_sessions(request: web.Request) -> web.Response:
-        query = request.query.get("q", "").strip()
-        if not query:
-            return web.json_response({"error": "q parameter required"}, status=400)
-        identity = getattr(request, "_api_identity", None)
-        is_admin = not identity or getattr(identity, "tier", "admin") == "admin"
-        limit = _safe_int_param(request, "limit", 20, hi=50)
-        channel_id = request.query.get("channel_id") or None
-        if not is_admin:
-            channel_id = identity.user_id
-        user_id = request.query.get("user_id") or None
-        after: float | None = None
-        before: float | None = None
-        if request.query.get("after"):
-            try:
-                after = float(request.query["after"])
-            except ValueError:
-                pass
-        if request.query.get("before"):
-            try:
-                before = float(request.query["before"])
-            except ValueError:
-                pass
-        results = await bot.sessions.search_history(
-            query, limit=limit, channel_id=channel_id,
-            user_id=user_id, after=after, before=before,
-        )
-        return web.json_response({"query": query, "results": results, "count": len(results)})
-
-    def _check_session_access(request: web.Request, channel_id: str) -> web.Response | None:
-        """Non-admin identities can only access their own session."""
-        identity = getattr(request, "_api_identity", None)
-        if not identity:
-            return None
-        if getattr(identity, "tier", "admin") == "admin":
-            return None
-        own_prefix = f"web:{identity.user_id}:session:"
-        if identity.user_id != channel_id and not channel_id.startswith(own_prefix):
-            return web.json_response({"error": "access denied"}, status=403)
-        return None
-
-    @routes.get("/api/sessions/{channel_id}")
-    async def get_session(request: web.Request) -> web.Response:
-        cid = request.match_info["channel_id"]
-        denied = _check_session_access(request, cid)
-        if denied:
-            return denied
-        session = bot.sessions.get(cid)
-        if not session:
-            return web.json_response({"error": "session not found"}, status=404)
-        messages = [
-            {
-                "role": m.role,
-                "content": m.content,
-                "timestamp": m.timestamp,
-                "user_id": m.user_id,
-            }
-            for m in session.messages
-        ]
-        return web.json_response({
-            "channel_id": cid,
-            "messages": messages,
-            "summary": session.summary,
-            "created_at": session.created_at,
-            "last_active": session.last_active,
-            "estimated_tokens": session.estimated_tokens,
-            "token_budget": bot.sessions.token_budget,
-        })
-
-    @routes.get("/api/sessions/{channel_id}/export")
-    async def export_session(request: web.Request) -> web.Response:
-        cid = request.match_info["channel_id"]
-        denied = _check_session_access(request, cid)
-        if denied:
-            return denied
-        session = bot.sessions.get(cid)
-        if not session:
-            return web.json_response({"error": "session not found"}, status=404)
-        fmt = request.query.get("format", "json")
-        messages = [
-            {
-                "role": m.role,
-                "content": m.content,
-                "timestamp": m.timestamp,
-                "user_id": m.user_id,
-            }
-            for m in session.messages
-        ]
-        safe_cid = _safe_filename(cid)
-        if fmt == "text":
-            lines = []
-            if session.summary:
-                lines.append(f"=== Summary ===\n{session.summary}\n")
-            lines.append(f"=== Messages ({len(messages)}) ===")
-            for m in messages:
-                ts = (
-                    time.strftime("%Y-%m-%d %H:%M:%S", time.localtime(m["timestamp"]))
-                    if m["timestamp"]
-                    else "?"
-                )
-                role = m["role"].upper()
-                uid = f" ({m['user_id']})" if m.get("user_id") else ""
-                lines.append(f"\n[{ts}] {role}{uid}:\n{m['content']}")
-            body = "\n".join(lines)
-            return web.Response(
-                text=body,
-                content_type="text/plain",
-                headers={"Content-Disposition": f'attachment; filename="session-{safe_cid}.txt"'},
-            )
-        # Default: JSON
-        export = {
-            "channel_id": cid,
-            "messages": messages,
-            "summary": session.summary,
-            "created_at": session.created_at,
-            "last_active": session.last_active,
-            "exported_at": time.time(),
-        }
-        return web.json_response(
-            export,
-            headers={"Content-Disposition": f'attachment; filename="session-{safe_cid}.json"'},
-        )
-
-    @routes.delete("/api/sessions/{channel_id}")
-    async def delete_session(request: web.Request) -> web.Response:
-        cid = request.match_info["channel_id"]
-        denied = _check_session_access(request, cid)
-        if denied:
-            return denied
-        if not bot.sessions.exists(cid):
-            return web.json_response({"error": "session not found"}, status=404)
-        bot.sessions.reset(cid)
-        return web.json_response({"status": "cleared"})
-
-    @routes.post("/api/sessions/clear-bulk")
-    async def clear_bulk_sessions(request: web.Request) -> web.Response:
-        denied = _require_admin(request)
-        if denied:
-            return denied
-        try:
-            data = await request.json()
-        except Exception:
-            return web.json_response({"error": "invalid JSON"}, status=400)
-        channel_ids = data.get("channel_ids", [])
-        if not isinstance(channel_ids, list) or not channel_ids:
-            return web.json_response(
-                {"error": "channel_ids must be a non-empty list"}, status=400
-            )
-        cleared = bot.sessions.reset_many(channel_ids)
-        return web.json_response({"status": "cleared", "count": cleared})
+    register_sessions(routes, bot)
 
     register_tools_meta(routes, bot)
 
@@ -1130,67 +799,7 @@ def create_api_routes(bot: OdinBot) -> web.RouteTableDef:
 
     register_aggregates(routes, bot)
 
-    # ------------------------------------------------------------------
-    # Trajectories
-    # ------------------------------------------------------------------
-
-    @routes.get("/api/trajectories")
-    async def list_trajectory_files(_request: web.Request) -> web.Response:
-        saver = getattr(bot, "trajectory_saver", None)
-        if saver is None:
-            return web.json_response({"error": "trajectory saving not available"}, status=503)
-        files = await saver.list_files()
-        return web.json_response({"files": files, "count": saver.count})
-
-    @routes.get("/api/trajectories/{filename}")
-    async def get_trajectory_file(request: web.Request) -> web.Response:
-        saver = getattr(bot, "trajectory_saver", None)
-        if saver is None:
-            return web.json_response({"error": "trajectory saving not available"}, status=503)
-        filename = request.match_info["filename"]
-        if (
-            not filename.endswith(".jsonl")
-            or "/" in filename
-            or "\\" in filename
-            or ".." in filename
-        ):
-            return web.json_response({"error": "invalid filename"}, status=400)
-        safe_path = (saver.directory / filename).resolve()
-        if not safe_path.is_relative_to(saver.directory.resolve()):
-            return web.json_response({"error": "invalid filename"}, status=400)
-        limit = _safe_int_param(request, "limit", 100, hi=500)
-        entries = await saver.read_file(filename, limit=limit)
-        return web.json_response({"entries": entries, "count": len(entries)})
-
-    @routes.get("/api/trajectories/message/{message_id}")
-    async def get_trajectory_by_message(request: web.Request) -> web.Response:
-        saver = getattr(bot, "trajectory_saver", None)
-        if saver is None:
-            return web.json_response({"error": "trajectory saving not available"}, status=503)
-        message_id = request.match_info["message_id"]
-        entry = await saver.find_by_message_id(message_id)
-        if entry is None:
-            return web.json_response({"error": "trajectory not found"}, status=404)
-        return web.json_response({"entry": entry})
-
-    @routes.get("/api/trajectories/search/query")
-    async def search_trajectories(request: web.Request) -> web.Response:
-        saver = getattr(bot, "trajectory_saver", None)
-        if saver is None:
-            return web.json_response({"error": "trajectory saving not available"}, status=503)
-        channel_id = request.query.get("channel_id")
-        user_id = request.query.get("user_id")
-        tool_name = request.query.get("tool_name")
-        errors_only = request.query.get("errors_only", "").lower() in ("1", "true")
-        limit = _safe_int_param(request, "limit", 50, hi=500)
-        results = await saver.search(
-            channel_id=channel_id,
-            user_id=user_id,
-            tool_name=tool_name,
-            errors_only=errors_only,
-            limit=limit,
-        )
-        return web.json_response({"results": results, "count": len(results)})
+    register_trajectories(routes, bot)
 
     register_skills(routes, bot)
 
@@ -1452,264 +1061,11 @@ def create_api_routes(bot: OdinBot) -> web.RouteTableDef:
 
     register_schedules(routes, bot)
 
-    # ------------------------------------------------------------------
-    # Autonomous loops
-    # ------------------------------------------------------------------
+    register_loops(routes, bot)
 
-    @routes.get("/api/loops")
-    async def list_loops(_request: web.Request) -> web.Response:
-        loops = []
-        for lid, info in bot.loop_manager._loops.items():
-            # Include last 5 iteration history entries
-            history = list(info._iteration_history)[-5:] if info._iteration_history else []
-            loops.append({
-                "id": lid,
-                "goal": info.goal,
-                "mode": info.mode,
-                "interval_seconds": info.interval_seconds,
-                "stop_condition": info.stop_condition,
-                "max_iterations": info.max_iterations,
-                "channel_id": info.channel_id,
-                "requester_id": info.requester_id,
-                "requester_name": info.requester_name,
-                "iteration_count": info.iteration_count,
-                "last_trigger": info.last_trigger,
-                "created_at": info.created_at,
-                "status": info.status,
-                "iteration_history": history,
-            })
-        return web.json_response(loops)
+    register_agents(routes, bot)
 
-    @routes.post("/api/loops")
-    async def start_loop(request: web.Request) -> web.Response:
-        data = await request.json()
-        goal = data.get("goal", "").strip()
-        if not goal:
-            return web.json_response({"error": "goal is required"}, status=400)
-        err = _validate_string(goal, "goal", _MAX_GOAL_LEN)
-        if err:
-            return web.json_response({"error": err}, status=400)
-        channel_id = data.get("channel_id", "").strip()
-        if not channel_id:
-            return web.json_response(
-                {"error": "channel_id is required"}, status=400
-            )
-        # Find the Discord channel to post to
-        try:
-            channel = bot.get_channel(int(channel_id))
-        except (ValueError, TypeError):
-            channel = None
-        if not channel:
-            return web.json_response({"error": "channel not found"}, status=404)
-
-        requester_id = "web-api"
-
-        # Build iteration callback (same pattern as _handle_start_loop)
-        async def _iteration_cb(
-            prompt: str, ch: object, prev_context: str | None,
-        ) -> str:
-            return await bot.tool_loop.run_autonomous(
-                prompt, ch, prev_context, requester_id,
-            )
-
-        result = bot.loop_manager.start_loop(
-            goal=goal,
-            channel=channel,
-            requester_id=requester_id,
-            requester_name="Web API",
-            iteration_callback=_iteration_cb,
-            interval_seconds=data.get("interval_seconds", 60),
-            mode=data.get("mode", "notify"),
-            stop_condition=data.get("stop_condition"),
-            max_iterations=data.get("max_iterations", 50),
-        )
-        if result.startswith("Error"):
-            return web.json_response({"error": result}, status=400)
-        return web.json_response({"loop_id": result}, status=201)
-
-    @routes.delete("/api/loops/{loop_id}")
-    async def stop_loop(request: web.Request) -> web.Response:
-        lid = request.match_info["loop_id"]
-        result = bot.loop_manager.stop_loop(lid)
-        is_error = "not found" in result.lower() or "not running" in result.lower()
-        return web.json_response(
-            {"result": result}, status=404 if is_error else 200
-        )
-
-    @routes.post("/api/loops/{loop_id}/restart")
-    async def restart_loop(request: web.Request) -> web.Response:
-        lid = request.match_info["loop_id"]
-        info = bot.loop_manager._loops.get(lid)
-        if not info:
-            return web.json_response({"error": "loop not found"}, status=404)
-
-        # Capture config before stopping
-        goal = info.goal
-        mode = info.mode
-        interval_seconds = info.interval_seconds
-        stop_condition = info.stop_condition
-        max_iterations = info.max_iterations
-        channel_id = info.channel_id
-        requester_id = info.requester_id
-        requester_name = info.requester_name
-
-        # Stop if running
-        if info.status == "running":
-            bot.loop_manager.stop_loop(lid)
-
-        # Find the channel
-        try:
-            channel = bot.get_channel(int(channel_id))
-        except (ValueError, TypeError):
-            channel = None
-        if not channel:
-            return web.json_response({"error": "channel not found"}, status=404)
-
-        # Build callback
-        async def _iteration_cb(
-            prompt: str, ch: object, prev_context: str | None,
-        ) -> str:
-            return await bot.tool_loop.run_autonomous(
-                prompt, ch, prev_context, requester_id,
-            )
-
-        new_id = bot.loop_manager.start_loop(
-            goal=goal,
-            channel=channel,
-            requester_id=requester_id,
-            requester_name=requester_name,
-            iteration_callback=_iteration_cb,
-            interval_seconds=interval_seconds,
-            mode=mode,
-            stop_condition=stop_condition,
-            max_iterations=max_iterations,
-        )
-        if new_id.startswith("Error"):
-            return web.json_response({"error": new_id}, status=400)
-        return web.json_response({"old_id": lid, "new_id": new_id}, status=201)
-
-    # ------------------------------------------------------------------
-    # Agents
-    # ------------------------------------------------------------------
-
-    @routes.get("/api/agents")
-    async def list_agents(_request: web.Request) -> web.Response:
-        try:
-            agent_agents = bot.agent_manager._agents
-            if not isinstance(agent_agents, dict):
-                return web.json_response([])
-        except (AttributeError, TypeError):
-            return web.json_response([])
-        agents = []
-        now = time.time()
-        for aid, info in agent_agents.items():
-            runtime = (info.ended_at or now) - info.created_at
-            agents.append({
-                "id": aid,
-                "label": info.label,
-                "goal": info.goal[:200],
-                "status": info.status,
-                "state": info.state.value if hasattr(info, "state") else info.status,
-                "channel_id": info.channel_id,
-                "requester_name": info.requester_name,
-                "iteration_count": info.iteration_count,
-                "tools_used": info.tools_used[-10:],
-                "runtime_seconds": round(runtime, 1),
-                "created_at": info.created_at,
-                "result": (info.result[:200] if info.result else ""),
-                "error": (info.error[:200] if info.error else ""),
-                "recovery_attempts": getattr(info, "recovery_attempts", 0),
-                "state_history": info._sm.history_as_dicts() if hasattr(info, "_sm") else [],
-                "depth": getattr(info, "depth", 0),
-                "parent_id": getattr(info, "parent_id", None),
-                "children_ids": list(getattr(info, "children_ids", [])),
-            })
-        return web.json_response(agents)
-
-    @routes.delete("/api/agents/{agent_id}")
-    async def kill_agent(request: web.Request) -> web.Response:
-        try:
-            if not isinstance(bot.agent_manager._agents, dict):
-                raise AttributeError
-        except (AttributeError, TypeError):
-            return web.json_response({"error": "no agent manager"}, status=404)
-        agent_id = request.match_info["agent_id"]
-        result = bot.agent_manager.kill(agent_id)
-        return web.json_response(
-            {"result": result}, status=404 if "not found" in result.lower() else 200
-        )
-
-    @routes.get("/api/agents/{agent_id}/children")
-    async def get_agent_children(request: web.Request) -> web.Response:
-        try:
-            mgr = bot.agent_manager
-        except (AttributeError, TypeError):
-            return web.json_response({"error": "no agent manager"}, status=503)
-        agent_id = request.match_info["agent_id"]
-        children = mgr.get_children(agent_id)
-        return web.json_response(children)
-
-    @routes.get("/api/agents/{agent_id}/lineage")
-    async def get_agent_lineage(request: web.Request) -> web.Response:
-        try:
-            mgr = bot.agent_manager
-        except (AttributeError, TypeError):
-            return web.json_response({"error": "no agent manager"}, status=503)
-        agent_id = request.match_info["agent_id"]
-        lineage = mgr.get_lineage(agent_id)
-        return web.json_response({"lineage": lineage})
-
-    @routes.get("/api/agents/{agent_id}/descendants")
-    async def get_agent_descendants(request: web.Request) -> web.Response:
-        try:
-            mgr = bot.agent_manager
-        except (AttributeError, TypeError):
-            return web.json_response({"error": "no agent manager"}, status=503)
-        agent_id = request.match_info["agent_id"]
-        descendants = mgr.get_descendants(agent_id)
-        return web.json_response({"descendants": descendants})
-
-    # ------------------------------------------------------------------
-    # Processes
-    # ------------------------------------------------------------------
-
-    @routes.get("/api/processes")
-    async def list_processes(_request: web.Request) -> web.Response:
-        registry = getattr(bot.tool_executor, "_process_registry", None)
-        if not registry:
-            return web.json_response([])
-        processes = []
-        now = time.time()
-        for pid, info in sorted(registry._processes.items()):
-            # Last 3 lines of output for inline preview
-            output_lines = list(info.output_buffer)
-            preview = [line.rstrip("\n") for line in output_lines[-3:]]
-            processes.append({
-                "pid": pid,
-                "command": info.command,
-                "host": info.host,
-                "status": info.status,
-                "exit_code": info.exit_code,
-                "uptime_seconds": round(now - info.start_time, 1),
-                "start_time": info.start_time,
-                "output_preview": preview,
-            })
-        return web.json_response(processes)
-
-    @routes.delete("/api/processes/{pid}")
-    async def kill_process(request: web.Request) -> web.Response:
-        registry = getattr(bot.tool_executor, "_process_registry", None)
-        if not registry:
-            return web.json_response({"error": "no process registry"}, status=404)
-        try:
-            pid = int(request.match_info["pid"])
-        except ValueError:
-            return web.json_response({"error": "invalid PID"}, status=400)
-        result = await registry.kill(pid)
-        is_error = "no process" in result.lower()
-        return web.json_response(
-            {"result": result}, status=404 if is_error else 200
-        )
+    register_processes(routes, bot)
 
     register_audit_log(routes, bot)
